ScalaFile.getClasses does not return fake classes anymore
Before, ScalaFile.getClasses returned all actually written, as well as
compiler generated fake classes of a file. This produced problems
when refactoring, because there getClasses is extensively used by
the idea code for moving files/packages/classes/...

To circumvent this problem a global variable `duringMoveRefactoring`
was introduced that controls whether the fake classse are returned as well.

It is not apparent that getClasses even has to return fake classes,
so this commit attempts to remove this altogether and makes getClasses
only return actually written classes.

SCL-12381